requirements.txt
Django==4.2.10
pillow
djangorestframework==3.15.2
Markdown==3.6

# API Dependencies
django-filter==24.2
drf-yasg==1.21.7
